+++ tubular/functions/strings.py
@@ -1,8 +1,10 @@
 """Stateless string transforms."""
 
 import narwhals as nw
+from beartype import beartype
 
 
+@beartype
 def convert_string_columns_to_lowercase(columns: list[str]) -> list[nw.Expr]:
     """Get expression for converting columns to lowercase.
 
@@ -14,6 +16,7 @@ def convert_string_columns_to_lowercase(columns: list[str]) -> list[nw.Expr]:
     return [nw.col(col).str.to_lowercase() for col in columns]
 
 
+@beartype
 def remove_characters_from_string_columns(
     columns: list[str], characters_formatted: str
 ) -> list[nw.Expr]:
@@ -25,3 +28,26 @@ def remove_characters_from_string_columns(
 
     """
     return [nw.col(col).str.replace_all(characters_formatted, "") for col in columns]
+
+
+@beartype
+def extract_string_components(
+    columns: list[str], by: str, return_n_components: int
+) -> list[nw.Expr]:
+    """Get expression for extracting components from a str columns, split by provided character.
+
+    Returns
+    -------
+    list[nw.Expr]: expressions for transformation
+
+    """
+    split_exprs = {col: nw.col(col).str.split(by=by) for col in columns}
+    return [
+        nw.when(split_exprs[col].list.len() > i)
+        .then(split_exprs[col])
+        .otherwise(None)
+        .list.get(i)
+        .alias(f"{col}_split_by_{by}_entry_{i}")
+        for col in columns
+        for i in range(return_n_components)
+    ]