+++ CHANGELOG.rst
@@ -35,6 +35,7 @@ Changed
 - added LowerCaseTransformer and RemoveCharactersTransformer to string module
 - added functions module for stateless transforms
 - removed narwhals private import from nominal module `#725 <https://github.com/azukds/tubular/issues/725>_`
+- setup get_transform_exprs methods for aggregations/comparison/dates/misc/numeric, and OneHotEncoder/MeanResponseTransformer
 
 3.3.0 (30/04/2026)
 ------------------